+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonGroup.cs
@@ -171,7 +171,7 @@ namespace Krypton.Toolkit
         [Category(@"Appearance")]
         [Description(@"The internal panel that contains group content.")]
         [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
-        public KryptonGroupPanel? Panel { get; }
+        public KryptonGroupPanel Panel { get; }
 
         /// <summary>
         /// Gets and sets the border style.